ClientWindow finally launches!
Tidied code up a bit.

Changed yet more Floats into Longs and fixed casting.

Replaced getName() returning this.name when extending Model, because this is lost when transferring objects as Model is not serializable. Instead we store this content locally in the class.

In Postcode and Supplier changed long (primitive) to Long (object), because as a primitive it does not work with the GUI provided by the University.

Changed Lists to ArrayLists just to be safe. This is because List is not serializable whereas ArrayList is serializable. While this might not be necessary in all cases, it's better to be safe than sorry!

Removed incorrectly placed notifyUpdate() that caused parts of the GUI to load on the client before this data had been read from the server yet, causing crashes due to NullPointers.

Implemented Listeners and notifyUpdate